pyproject.toml
[project]
name = "scopesim-targets"
version = "0.1.0a1"
description = "Definition of on-sky targets for the ScopeSim telescope observation simulator."
authors = [
    {name = "Fabian Haberhauer",email = "fabian.haberhauer@univie.ac.at"}
]
license = {text = "GPL-3.0-or-later"}
readme = "README.md"
requires-python = ">=3.10, <3.14"
dependencies = [
    "numpy (>=1.26.4,<2.3.0)",
    "scipy (>=1.15.3,<1.16.0)",
    "astropy (>=6.1.7,<8.0.0)",
    "matplotlib (>=3.10.1,<3.11.0)",
    "pyyaml (>=6.0.2,<7.0.0)",

    "synphot (>=1.6.0,<2.0.0)",
    "astar-utils (>=0.3.3)"
]


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.8"
black = "^25.1.0"
mypy = "^1.17.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.4.2"
pytest-cov = "^6.1.1"


[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^8.1.3"
sphinx-book-theme = "^1.1.3"
sphinx-copybutton = "^0.5.2"
sphinxcontrib-apidoc = ">=0.6.0"
sphinxcontrib-mermaid = "^1.0.0"
myst-nb = "^1.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
omit = ["tests/*", "docs/*"]
